Change cursor colour in "iTerm2 Default" colour scheme
This is because Windows Terminal does not allow setting the cursor text
colour, that is, the colour of the character under the cursor (the text
under the cursor just stays in its native colour). This may make the
character under the cursor very illegible if a bright text colour is
used.

For this reason, the cursor colour is changed to darker grey so that the
contrast with bright text is better.

A feature that allows the cursor text colour to be changed is currently
under developement and will probably be included in a future release of
Windows Terminal (see comments starting from [1]).

When this feature will be included in Windows Terminal, the change in
this commit can be reverted.

[1] https://github.com/microsoft/terminal/issues/1203#issuecomment-1652269433