Restore original cursor colour in iTerm2 Default colour scheme
The preview version of Windows Terminal v1.18 fixes the issue of
illegible text under the cursor by somehow automatically determining the
colour of the character under the cursor.

The colour of the character under the cursor is not constant, it changes
depending on the native colour of the character (and probably the colour
of the cursor). However, the contrast is acceptable in most cases. See
more information about this feature in [1].

Setting the cursor text colour explicitly as part of the colour scheme
(arguably a better solution than the current implementation) is still
separately discussed in [2].

[1] https://github.com/microsoft/terminal/issues/7118
[2] https://github.com/microsoft/terminal/issues/15766